==> 1-prepare_data.Rmd <==
---
title: "1-prepare_data.Rmd"
author: "Antton Alberdi"
date: "2024-01-13"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load libraries

```{r libraries, warning=FALSE, comments="", message=FALSE}
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(ape))
suppressPackageStartupMessages(library(distillR))
```

## Retrieve and load data
Raw data files are stored at the University of Copenhagen's Electronic Research Data Archive (ERDA). Shared links must be created in order to retrieve the files directly by R, as shown in the scripts below.  

### Count table
This is the document containing the number of sequencing reads from each sample have been mapped to each MAG. Note that this is the raw data that needs to be further processed before running any statistics on them.

```{r load_count, message=FALSE}
read_counts <- read_tsv("data/genome.count.tsv") %>%
  rename_all(~ str_remove_all(., ".lib1")) %>% #simplify column names
  rename(genome = 1)
```

Generate a vector of genome names to be employed for filtering and sorting data in downstream steps.
```{r unit_list, message=FALSE}
genomes <- read_counts$genome # create list of genome names
samples <- read_counts %>% select(-genome) %>% colnames() # create list of genome names

```

### Base hit table
This is the document containing the number of nucleotide bases have been covered by at least one read in each sample and MAG. This information is used to calculate MAG coverage values.

```{r load_hits, message=FALSE}
basehits <- read_tsv("data/genome.covered_bases.tsv") %>%
  rename_all(~ str_remove_all(., ".lib1")) %>% #simplify column names
  rename(genome = 1)
```

### Sample metadata

```{r load_sample_metadata, message=FALSE}
sample_metadata <- read_csv("data/sample_metadata.csv") %>%
  filter(sample %in% samples)
```

### Genome metadata
Relevant metadata of genomes is fetched from 2-3 files and merged into one genome metadata object for downstream analyses.

#### Taxonomy
This is the raw taxonomy table generated by GTDBtk, which is simplified for downstream analyses.
```{r load_taxonomy, message=FALSE}
genome_taxonomy <- read_tsv("data/gtdbtk.summary.tsv") %>%
  rename(genome = user_genome) %>%
  mutate(genome = str_replace_all(genome,"\\.fa", "")) %>%
  separate(classification, c("domain","phylum","class","order","family","genus","species"),  sep =";") %>%
  select(genome,domain,phylum,class,order,family,genus,species) %>%
  arrange(match(genome, genomes))
```

#### Genome lengths
Number of nucleotides of each genome. This information is needed to calculate coverage values and to normalise read counts by genome length.
```{r load_lengths, message=FALSE}
genome_lengths <- read_tsv("data/genome.length.tsv") %>%
  rename(genome = 1, length=2) %>% 
  select(genome,length) %>% #remove redundancy
  arrange(match(genome, genomes))
```

#### Genome quality
Quality properties of the genomes. 
```{r load_quality, message=FALSE}
genome_quality <- read_tsv("data/quality_report.tsv") %>%
  rename(genome = 1) %>%
  arrange(match(genome, genomes)) %>%
  select(genome, Completeness, Contamination, Coding_Density, Genome_Size) %>%
  rename(completeness=Completeness,contamination=Contamination,coding_density=Coding_Density,length=Genome_Size)
```

#### Merged metadata object
Merge taxonomy, length and quality information
```{r create_genomemetadata, message=FALSE}
genome_metadata <- genome_taxonomy %>%
  left_join(genome_quality,by=join_by(genome==genome)) #join quality
```

### Genome tree
This is the raw tree generated by GTDBtk, which needs to be pruned to obtain the phylogenetic tree of the genomes. Note that the archaeal tree is only generated if any archaeans are detected among the genomes.
```{r load_tree, message=FALSE, warning=FALSE}
genome_tree <- read.tree("data/gtdbtk.backbone.bac120.classify.tree")
genome_tree$tip.label <- str_replace_all(genome_tree$tip.label,"'", "") #remove single quotes in MAG names
genome_tree <- keep.tip(genome_tree, tip=genomes) # keep only MAG tips
```

### MAG functional annotations
This is the raw annotation table generated by DRAM, which is used to generate GIFT data using distillR.
```{r load_annotations, message=FALSE}
genome_annotations <- read_tsv("data/annotations.tsv.gz") %>%
  rename(gene=1,genome=2)
```

## Filter and normalise data
Raw data needs to be filtered and normalised to make it useful for downstream analyses. 

### Generate coverage table
By dividing the number of base hits by the length of each genome, coverage values can be calculated.

```{r calc_coverage}
genome_coverage <- basehits %>%
  mutate(across(where(is.numeric), ~ ./genome_metadata$length))
```

### Coverage filtering
Genomes that have less than 30% of their length covered by reads are turned into zeros to account for the random allocation of reads across genomes due to mapping heuristics. 

```{r filter_coverage}
min_coverage=0.3
read_counts_filt <- genome_coverage %>%
  mutate(across(where(is.numeric), ~ ifelse(. > min_coverage, 1, 0))) %>%
  mutate(across(-1, ~ . * read_counts[[cur_column()]])) 
```

### Generate genome count table
After filtering the low-coverage reads, read counts are transformed into genome counts using genome-length and read-length information.

```{r calc_genometable}
readlength=150 #change if sequencing read length is different
genome_counts <- read_counts_filt %>%
  mutate(across(where(is.numeric), ~ . / (genome_metadata$length / readlength) ))
```

### Distil functional annotations
Raw functional annotations are distilled into genome-inferred functional traits to generate biologically more meaningful functional traits for downstream analyses.

```{r distill_annotations, warning=FALSE, comments="", message=FALSE, results='hide'}
genome_gifts <- distill(genome_annotations,GIFT_db,genomecol=2,annotcol=c(9,10,19))
```

## Wrap working objects
In the last step, the objects that are needed for downstream analyses are stored in an R object.

```{r wrap_objects}
save(read_counts, genome_counts, genome_tree, genome_metadata, genome_gifts, sample_metadata, file = "data/data.Rdata")
```

- **read_counts**: Number of reads mapped to each genome in each sample. Note this is the unfiltered and unnormalised raw community composition table.
- **genome_counts**: Number of genomes quantified in each sample, calculated through filtering and normalising ***read_counts***. This is the community composition table to be used in downstream analyses unless otherwise stated.
- **genome_tree**: Phylogenetic tree of the genomes, to be employed in downstream phylogenetic analyses.
- **genome_metadata**: Taxonomic and quality information of the genomes.
- **genome_gifts**: Genome-inferred functional traits of the genomes, to be employed in downstream functional analyses.
- **sample_metadata**: Treatment/population and other relevant metadata of the samples.
